==> models/saiva-3-day-hosp-v5/src/notebooks/08-model-config.ipynb <==
### Create trained models in db

import sys
import json

from omegaconf import OmegaConf

sys.path.insert(0, '/src')

## Load config

from shared.constants import saiva_api, LOCAL_TRAINING_CONFIG_PATH
from shared.utils import load_config

config = load_config(LOCAL_TRAINING_CONFIG_PATH)
training_config = config.training_config

with open('model_config.json', 'r') as file:
    model_config = json.load(file)

training_metadata = training_config.training_metadata
training_metadata['ml_model_org_config_ids'] = [config.id for config in training_config.ml_model_org_configs]

for individual_model_config in model_config:
    test_recall_at_rank_15 = individual_model_config['test_recall_at_rank_15']
    test_recall_at_rank_15 = test_recall_at_rank_15 if test_recall_at_rank_15 != 'nan' else None
    individual_model_config['test_recall_at_rank_15'] = test_recall_at_rank_15
    saiva_api.trained_models.create_with_dict(ml_model_org_config_id=None, training_metadata=OmegaConf.to_container(training_metadata, resolve=True), model_config=individual_model_config)
